requirements.txt
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.20.0
packaging==20.9
pathy==0.3.4
preshed==3.0.5
pydantic==1.7.3
pyparsing==2.4.7
python-Levenshtein==0.12.2
PyYAML==5.4.1
requests==2.25.1
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
srsly==2.4.0
thinc==8.0.1
tqdm==4.56.0
typer==0.3.2
urllib3==1.26.3
wasabi==0.8.2
